-i/-x OPTIONS SUPPORT PATTERN FORM '<MODULE-REGEX>!<FUNCTION-REGEX>'
Currently the -i (IncludeFunction) and -x (ExcludeFunction) command
line options take a function regex pattern, but will match functions
in *ALL* modules in the process.  There is no easy way to request a
function regex pattern in a *single* module.

This commit enhances the -i and -x function options to support patterns
that specify both a module *and* a function regex.

Examples:

-i '*free*'
        Matches all functions with 'free' in its name in ALL modules
	Equivalent to 'exports:*!*free*'

-i 'msvcrt.dll!*cpy*'
        Matches all functions with 'cpy' only in msvcrt.dll
	Equivalent to 'exports:msvcrt.dll!*cpy*'

-i 'msvcrt.dll!'
        Equivalent to 'exports:msvcrt.dll:*'

-i '!*cpy*'
        Equivalent to 'exports:*!*cpy*'