INITIALIZE SESSION 'STATE' VARIABLE VIA JS SCRIPTS
This new feature enables a user to initialize the 'state' global
variable before starting a trace session.  This is done by specifying
any number of Javascript files via a command line option
('-k' / '--init-session').  These files will be executed before the
tracing session begins.  This means that the `state` global variable,
passed to every handler, can be pre-populated with functions and
variables.  This opens the door to shared handler library code and
variables.

The feature implementation pivots on executing the initialize session
scripts using Javascript's 'eval' function.  It turns out that eval()
has its own idiosyncracies, which means the developer needs to adhere
to some basic code writing rules (see below).

Unconnected to the existence of any session initialization scripts,
tracer.py now creates three pre-defined objects in the 'state' object
before beginning the trace:

    // Create namespaces for user functions ('F') and data ('D'),
    pre-populate if relevant
    state.F = {}
    state.D = {}
    state.D.NEWLINE = String.fromCharCode (10);

Caveat: Use Standard Javascript Syntax
======================================

The Javascript code written in session initializing scripts should
conform to Standard Javascript syntax conform to Standard Javascript
syntax:

- Insert semicolons at the end of all statements as required in
  standard Javascript
- Define all variables using the 'var' keyword.

Caveat: Do not embed the characters '\n' in the script
======================================================

The eval() function acts strangely in the presence of newlines,
returning an error if encountered:

  - Embedded newline (0x0A) characters, i.e., multiple JS lines
  - The occurrence of `\n' in a string

The feature as implemented converts 0x0A characters to the Unicode
'\u000a' before passing the JS to the eval() function, so the user
does not need to do anything special here.

The occurrence of '\n' in the script is a problem for eval(), and will
result in an 'unterminated string' error message.  (this explains why
all examples of eval() on the web are limited to a single line).  To
solve this, the developer should substituted all occurrences of '\n'
with String.fromCharCode(10).  As mentioned above, there is a
predefined state.D.NEWLINE string that can be used wherever '\n' is
needed.

Thus, printing a string in an initialize session script should be done
as follows:

    state.F.bar = function (log) {
        log ('BAR function called' + state.D.NEWLINE);
    }

rather than:

    state.F.bar = function (log, msg) {
        log ('BAR function called\n');
    }