Add -O/--options-file for passing args via file
To enable a user to pass command line arguments both on the command line
and via one or more text files.  This is done by specifying the -O or
the --options-file command line option, followed by a path to a text
file containing additional command line arguments.  Within the text
file(s), arguments can be on the same or different lines, delimited by
white space or newlines.  The arguments found in the text files are
tokenized correctly (using the shlex Python package) and eventually
passed to parser.parse_args() for processing.

The command line and/or an argument file can include any number of -O /
--options-file arguments.  The feature will process them all in their
correct order until all arguments have been tokenized.  The feature
will flag an error if an argument file is specified more than once (to
prevent endless looping).

For example, supposing the file '/mnt/d/test/more-args.txt' contains
the following three lines, showing arguments delimited by both spaces
and newlines:

    -U
    -i 'libfoo.so!*mem*' -i '*free*'
    --decorate -I 'libbar.so'

A command like:

    frida-trace --quiet -O /mnt/d/test/more-args.txt com.jumangi.pq

will be equivalent to the following command (all arguments on one line):

    frida-trace --quiet -U -i 'libfoo.so!*mem*' -i '*free*' --decorate
                -I 'libbar.so' com.jumangi.pq

The rationale for this feature is when one needs to specify hundreds or
even thousands of -i / -x / -a options and values in a single run.  Some
operating systems limit the size of a command line, leaving the user no
way to pass the arguments.