POPULATE <STATE> WITH USER DATA BEFORE SESSION
Enable more dynamic control of handlers by allowing the user to pass
'user data' via the command line.  This user data is assigned to the
'state.userdata' string variable and can be access by a handler during
run time.  Based on its value, the user can control the runtime
behaviour of handlers.

In its simplest form, the user data string can be a string, e.g.:

    frida-trace ... -u 'depth=5'

A recommended format of user data is to use a robust JSON structure that
can be used to pass different data to different handlers.  Here is a
more sophisticated JSON format:

    {
        "trace_data" : {
                           "num_funcs": 25,
	                   "max_depth": 4
                       },
        "memcpy_params" : {
                              "max_bytes_copied": 512,
	                      "is_higher_address": true
                          }
    }

This should be compressed into a string without newlines and unnecessary
spaces (on one single line):

    frida-trace ...
        -u '{"trace_data":{"num_funcs":25,"max_depth":4},
	     "memcpy_params":{"max_bytes_copied":512,"is_higher_address":true}}'

A handler can pass the string in <state.userdata> to JSON.parse() to
access anything it needs.