Use default handler for SIGPIPE
Python override the SIGPIPE signal handler with its own handler. Ideally
we would leave the default handler and catch the raised exception
(BrokenPipeError) but there's no way to supress the unhandled error
completly. So instead of catching the exception we prevent it from being
raised by using the default handler.