application: Use default handler for SIGPIPE (#56)
Python overrides the SIGPIPE signal handler with its own handler.
Ideally we would leave the default handler and catch the raised
exception (BrokenPipeError) but there's no way to suppress the unhandled
error completely. So instead of catching the exception we prevent it
from being raised by using the default handler.