Use correct directory for each file type
The current implementation did not fit well in any operation system.
There was some work on support xdg directories but in reality it was
rarely used because most people use the default XDG directory without
setting the environment variable.

This patch support XDG directories properly in addition of supporting
the native folder location in Windows (someone with a bit more knowledge
about MacOS could improve it there too). Not only that but I made sure
to separate the directories to config, state and data directories just
like in the XDG specification (there are more directories that were
ignored for now).