Migrate from deprecated optparse to argparse
Optparse is deprecated since Python 3.2 (more than 10 years ago!) so
it's time to migrate to the newer argparse module that is still
supported.

There wasn't much to change because argparse tries as best as it can to
be backward compatible. Mostly what was changed is callback arguments
and handling of extra arguments.

I also deleted redundant arguments that were the default anyway.