repl: Use correct directory for each file type (#73)
The current implementation did not fit well in any operating system.
There was some work on supporting XDG directories but in reality it was
rarely used because most people use the default XDG directory without
setting the environment variable.

This patch support XDG directories properly in addition to supporting
the native directory location on Windows (someone with a bit more
knowledge about macOS could improve it there too). Not only that but I
made sure to separate the directories to config, state and data
directories just like in the XDG specification (there are more
directories that were ignored for now).