apk: Update ResourceMap, insert debuggable at the correct position
This commit is the result of an investigation into why `frida-apk`
failed on a stock Pixel 3 with Android 12. Much of the explanation
assumes knowledge of the Android Binary XML format, which to a large
extent is documented at
https://justanapplication.wordpress.com/category/android/android-binary-xml/.

First, in order to recognize an app to be debuggable, its resource ID
needs to be specified in the resource map chunk. While `ResourceTypes.h`
claims that the resource map chunk would be optional, that's not the
case for manifests in practice. This means we first look out for the
matching chunk, and pad it to insert the debuggable resource ID (string
pool and resource map share the same indices).

Second, for some reasons the implementation on my test device expects
tag attributes to be sorted by their resource ID. It's not really clear
why this is the case and I have not managed to turn up any sources that
indicate such processing, but moving `debuggable` to the front of the
attribute list fixes the behavior. However, that then breaks all
attributes with lower resource IDs listed afterwards, so we actually
need to traverse the attribute list to figure out where we need to
insert.

Fixes #96.