Add --kill-on-exit to repl
This option makes Frida kill the spawned process (if any)
when Frida exists. This is useful for running Frida from another
program - sending a SIGKILL (or the platform equivalent) to Frida
can be used to kill the spawned program as well, without the need
to determine the pid of the spawned program from the parent program.